+++ lib/vagrant/errors.rb
@@ -60,7 +60,7 @@ module Vagrant
         message = { :_key => message } if message && !message.is_a?(Hash)
         message = { :_key => error_key, :_namespace => error_namespace }.merge(message || {})
 
-        if error_key
+        if message[:_key]
           message = translate_error(message)
         else
           message = error_message